Extract view changer node-related bindings
