+++ plenum/server/node.py
@@ -1,5 +1,4 @@
 import json
-import logging
 import os
 import random
 import time
@@ -22,7 +21,7 @@ from plenum.server.backup_instance_faulty_processor import BackupInstanceFaultyP
 from plenum.server.inconsistency_watchers import NetworkInconsistencyWatcher
 from plenum.server.last_sent_pp_store_helper import LastSentPpStoreHelper
 from plenum.server.quota_control import StaticQuotaControl, RequestQueueQuotaControl
-from plenum.server.view_change.pre_view_change_strategies import preVCStrategies
+from plenum.server.view_change.node_view_changer import create_view_changer
 from state.pruning_state import PruningState
 from state.state import State
 from storage.helper import initKeyValueStorage, initHashStore, initKeyValueStorageIntKeys
@@ -118,7 +117,7 @@ from plenum.server.req_handler import RequestHandler
 from plenum.server.router import Router
 from plenum.server.suspicion_codes import Suspicions
 from plenum.server.validator_info_tool import ValidatorNodeInfoTool
-from plenum.server.view_change.view_changer import ViewChanger, ViewChangerDataProvider
+from plenum.server.view_change.view_changer import ViewChanger
 
 pluginManager = PluginManager()
 logger = getlogger()
@@ -216,89 +215,6 @@ class GcTimeTracker:
             elapsed = time.perf_counter() - start
             self._metrics.add_event(MetricsName.GC_GEN0_TIME + gen, elapsed)
             self._timestamps[gen] = None
-
-
-class ViewChangerNodeDataProvider(ViewChangerDataProvider):
-    def __init__(self, node):
-        self._node = node
-
-    def name(self) -> str:
-        return self._node.name
-
-    def config(self) -> object:
-        return self._node.config
-
-    def quorums(self) -> Quorums:
-        return self._node.quorums
-
-    def has_pool_ledger(self) -> bool:
-        return self._node.poolLedger is not None
-
-    def ledger_summary(self) -> List[Tuple[int, int, str]]:
-        return self._node.ledger_summary
-
-    def node_registry(self, size):
-        return self._node.poolManager.getNodeRegistry(size)
-
-    def is_node_synced(self) -> bool:
-        return self._node.is_synced
-
-    def node_mode(self) -> Mode:
-        return self._node.mode
-
-    def next_primary_name(self) -> str:
-        return self._node.elector._next_primary_node_name_for_master()
-
-    def current_primary_name(self) -> str:
-        return self._node.master_primary_name
-
-    def has_primary(self) -> bool:
-        return self._node.master_replica.hasPrimary
-
-    def is_primary(self):
-        return self._node.master_replica.isPrimary
-
-    def is_primary_disconnected(self) -> bool:
-        return \
-            self._node.primaries_disconnection_times[self._node.master_replica.instId] \
-            and self._node.master_primary_name \
-            and self._node.master_primary_name not in self._node.nodestack.conns
-
-    def is_master_degraded(self) -> bool:
-        return self._node.monitor.isMasterDegraded()
-
-    def pretty_metrics(self) -> str:
-        return self._node.monitor.prettymetrics
-
-    def state_freshness(self) -> float:
-        replica = self._node.master_replica
-        timestamps = replica.get_ledgers_last_update_time().values()
-        oldest_timestamp = min(timestamps)
-        return replica.get_time_for_3pc_batch() - oldest_timestamp
-
-    def connected_nodes(self) -> Set[str]:
-        return self._node.nodestack.connecteds
-
-    def notify_view_change_start(self):
-        self._node.on_view_change_start()
-
-    def notify_view_change_complete(self):
-        self._node.on_view_change_complete()
-
-    def notify_initial_propose_view_change(self):
-        self._node.schedule_initial_propose_view_change()
-
-    def start_catchup(self):
-        self._node.start_catchup()
-
-    def restore_backup_replicas(self):
-        self._node.backup_instance_faulty_processor.restore_replicas()
-
-    def select_primaries(self, node_reg):
-        self._node.select_primaries(node_reg)
-
-    def discard(self, msg, reason, logMethod=logging.error, cliOutput=False):
-        self._node.discard(msg, reason, logMethod, cliOutput)
 
 
 class Node(HasActionQueue, Motor, Propagator, MessageProcessor, HasFileStorage,
@@ -1334,10 +1250,7 @@ class Node(HasActionQueue, Motor, Propagator, MessageProcessor, HasFileStorage,
         if self.view_changer:
             return self.view_changer
         else:
-            vc = ViewChanger(ViewChangerNodeDataProvider(self))
-            if hasattr(self.config, 'PRE_VC_STRATEGY'):
-                vc.pre_vc_strategy = preVCStrategies.get(self.config.PRE_VC_STRATEGY)(vc, self)
-            return vc
+            return create_view_changer(self)
 
     def newPrimaryDecider(self):
         if self.primaryDecider:
