+++ plenum/test/primary_selection/test_primary_selector.py
@@ -8,14 +8,14 @@ from plenum.common.util import get_utc_epoch
 
 from plenum.server.propagator import Requests
 
-from plenum.server.node import Node, ViewChangerNodeDataProvider
+from plenum.server.node import Node
 
 from plenum.common.metrics_collector import NullMetricsCollector
+from plenum.server.view_change.node_view_changer import create_view_changer
 from stp_core.types import HA
 
 from plenum.common.startable import Mode
 from plenum.server.primary_selector import PrimarySelector
-from plenum.server.view_change.view_changer import ViewChanger
 from plenum.common.messages.node_messages import ViewChangeDone
 from plenum.server.quorums import Quorums
 from plenum.server.replica import Replica
@@ -66,7 +66,7 @@ class FakeNode:
         self.ledgerManager.addLedger(0, ledger0)
         self.ledgerManager.addLedger(1, ledger1)
         self.quorums = Quorums(self.totalNodes)
-        self.view_changer = ViewChanger(ViewChangerNodeDataProvider(self))
+        self.view_changer = create_view_changer(self)
         self.elector = PrimarySelector(self)
         self.metrics = NullMetricsCollector()
