+++ plenum/test/view_change/conftest.py
@@ -4,9 +4,9 @@ from collections import deque
 import pytest
 
 from plenum.common.util import get_utc_epoch
-from plenum.server.node import Node, ViewChangerNodeDataProvider
+from plenum.server.node import Node
 from plenum.server.quorums import Quorums
-from plenum.server.view_change.view_changer import ViewChanger
+from plenum.server.view_change.node_view_changer import create_view_changer
 from plenum.test.conftest import getValueFromModule
 from plenum.test.primary_selection.test_primary_selector import FakeNode
 from plenum.test.test_node import getRequiredInstances
@@ -60,7 +60,7 @@ def fake_view_changer(request, tconf):
         master_primary_name='Alpha',
         master_replica=FakeSomething(instId=0)
     )
-    view_changer = ViewChanger(ViewChangerNodeDataProvider(node))
+    view_changer = create_view_changer(node)
     # TODO: This is a hack, do something better
     view_changer.node = node
     return view_changer
